docs(aio): ward's changes
